Pass through start.elf bootargs on Raspberry Pi
On Raspberry Pi, the primary bootloader start.elf uses the options in
config.txt, as well as options hidden in the firmware itself, to tell
the Linux kernel e.g. framebuffer sizes, memory regions, MAC addresses
and more.

Normally, u-boot would not be able to pass through these options to
the Linux kernel. This patch adds pass-through support via an
additional env variable "bootargs_orig" which is loaded with the
original cmdline obtained from ATAG or the FDT.

In addition, this allows the user to configure DT overlays the "Pi
way" by enabling them in config.txt, as this patch passes through the
full FDT if it has been passed one by u-boot.

Note that, as stated in the comment block, for FDT passthrough to
work, the u-boot.img must be run through a perl script by the
Raspberry Pi Foundation, which appends a trailer to the image telling
start.elf to supply a FDT instead of an ATAG struct.

Signed-off-by: Marco Schuster <m.schuster91@googlemail.com>