Remove fdtfile list from Raspberry Pi
As pass-through of the processed FDT is working now, there is no need
for the RPi u-boot any more to load fdt files. This patch removes the
fdtfile list and the setter for the env variable "fdtfile".

This also removes the naming inconsistencies between the fdtfile names
in u-boot and the fdtfile names in the official Raspberry Pi  Firmware
repository.

Signed-off-by: Marco Schuster <m.schuster91@googlemail.com>